pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "surya-tabular-ocr"
version = "0.17.1"
description = "English OCR, layout analysis, and table recognition from document images"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
authors = [{ name = "Saksham Saxena" }]
keywords = ["ocr", "pdf", "table recognition", "text detection", "layout analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Typing :: Typed",
]
dependencies = [
    "transformers>=4.56.1,<5",
    "torch>=2.7.0,<3",
    "pydantic>=2.5.3,<3",
    "pydantic-settings>=2.1.0,<3",
    "python-dotenv>=1.0.0,<2",
    "pillow>=10.2.0",

    "platformdirs>=4.3.6,<5",
    "opencv-python-headless==4.11.0.86",
    "einops>=0.8.1,<1",
]

[project.urls]
Repository = "https://github.com/sakshamsaxena/surya"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pre-commit>=4.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["surya"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::DeprecationWarning",
]
